+++ aa_forum/views/admin.py
@@ -12,7 +12,7 @@ from django.utils.safestring import mark_safe
 from django.utils.translation import gettext as _
 
 from aa_forum.forms import EditBoardForm, EditCategoryForm
-from aa_forum.models import Boards, Categories
+from aa_forum.models import Boards, Categories, Topics
 
 
 @login_required
@@ -305,3 +305,40 @@ def ajax_board_order(request: WSGIRequest) -> JsonResponse:
         data.append({"success": True})
 
     return JsonResponse(data, safe=False)
+
+
+@login_required
+@permission_required("aa_forum.manage_forum")
+def topic_change_lock_state(
+    request: WSGIRequest, topic_id: int
+) -> HttpResponseRedirect:
+    """
+    Change the lock state of the given topic
+    :param request:
+    :type request:
+    :param topic_id:
+    :type topic_id:
+    :return:
+    :rtype:
+    """
+
+    topic = Topics.objects.get(pk=topic_id)
+
+    if topic.is_locked:
+        topic.is_locked = False
+
+        messages.success(
+            request,
+            mark_safe(_("<h4>Success!</h4><p>Topic has been unlocked/re-opened.</p>")),
+        )
+    else:
+        topic.is_locked = True
+
+        messages.success(
+            request,
+            mark_safe(_("<h4>Success!</h4><p>Topic has been locked/closed.</p>")),
+        )
+
+    topic.save()
+
+    return redirect("aa_forum:forum_board", topic.board.category.slug, topic.board.slug)